+++ lib/core/domain/extension/future_extensions.dart
