+++ lib/core/domain/utils/console_output.dart
