+++ lib/sdk/sdk_apple_health/ios_configuration.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
-import 'package:rook_flutter_sdk/common/console_output.dart';
-import 'package:rook_flutter_sdk/common/environments.dart';
-import 'package:rook_flutter_sdk/common/widget/scrollable_scaffold.dart';
-import 'package:rook_flutter_sdk/common/widget/section_title.dart';
-import 'package:rook_flutter_sdk/features/sdk_apple_health/ios_background_sync.dart';
-import 'package:rook_flutter_sdk/features/sdk_apple_health/ios_continuous_upload.dart';
-import 'package:rook_flutter_sdk/features/sdk_apple_health/ios_sync.dart';
-import 'package:rook_flutter_sdk/features/sdk_apple_health/ios_user_management.dart';
+import 'package:rook_flutter_sdk/core/domain/utils/console_output.dart';
+import 'package:rook_flutter_sdk/core/domain/utils/environments.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/scrollable_scaffold.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/section_title.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_apple_health/ios_background_sync.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_apple_health/ios_continuous_upload.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_apple_health/ios_sync.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_apple_health/ios_user_management.dart';
 import 'package:rook_flutter_sdk/secrets.dart';
 import 'package:rook_sdk_apple_health/rook_sdk_apple_health.dart';
 import 'package:rook_sdk_core/rook_sdk_core.dart';
