+++ lib/sdk/sdk_health_connect/android_configuration.dart
@@ -1,28 +1,29 @@
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
-import 'package:rook_flutter_sdk/common/console_output.dart';
-import 'package:rook_flutter_sdk/common/environments.dart';
-import 'package:rook_flutter_sdk/common/widget/scrollable_scaffold.dart';
-import 'package:rook_flutter_sdk/common/widget/section_title.dart';
-import 'package:rook_flutter_sdk/features/sdk_samsung_health/samsung_background_sync.dart';
-import 'package:rook_flutter_sdk/features/sdk_samsung_health/samsung_permissions.dart';
-import 'package:rook_flutter_sdk/features/sdk_samsung_health/samsung_sync.dart';
-import 'package:rook_flutter_sdk/features/sdk_samsung_health/samsung_user_management.dart';
+import 'package:rook_flutter_sdk/core/domain/utils/console_output.dart';
+import 'package:rook_flutter_sdk/core/domain/utils/environments.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/scrollable_scaffold.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/section_title.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_health_connect/android_background_steps.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_health_connect/android_background_sync.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_health_connect/android_permissions.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_health_connect/android_sync.dart';
+import 'package:rook_flutter_sdk/sdk/sdk_health_connect/android_user_management.dart';
 import 'package:rook_flutter_sdk/secrets.dart';
 import 'package:rook_sdk_core/rook_sdk_core.dart';
-import 'package:rook_sdk_samsung_health/rook_sdk_samsung_health.dart';
+import 'package:rook_sdk_health_connect/rook_sdk_health_connect.dart';
 
-const String samsungConfigurationRoute = '/samsung/configuration';
+const String androidConfigurationRoute = '/android/configuration';
 
-class SamsungConfiguration extends StatefulWidget {
-  const SamsungConfiguration({super.key});
+class AndroidConfiguration extends StatefulWidget {
+  const AndroidConfiguration({super.key});
 
   @override
-  State<SamsungConfiguration> createState() => _SamsungConfigurationState();
+  State<AndroidConfiguration> createState() => _AndroidConfigurationState();
 }
 
-class _SamsungConfigurationState extends State<SamsungConfiguration> {
-  final Logger logger = Logger('SamsungConfiguration');
+class _AndroidConfigurationState extends State<AndroidConfiguration> {
+  final Logger logger = Logger('AndroidConfiguration');
 
   final ConsoleOutput configurationOutput = ConsoleOutput();
   final ConsoleOutput initializeOutput = ConsoleOutput();
@@ -39,11 +40,16 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
       name: 'SDK Configuration',
       child: Column(
         children: [
-          const SectionTitle('1. Initialize SDK'),
+          const SectionTitle('1. Configure SDK'),
           Text(configurationOutput.current),
+          FilledButton(
+            onPressed: setConfiguration,
+            child: const Text('setConfiguration'),
+          ),
+          const SectionTitle('2. Initialize SDK'),
           Text(initializeOutput.current),
           FilledButton(onPressed: initialize, child: const Text('initRook')),
-          const SectionTitle('2. Update user ID'),
+          const SectionTitle('3. Update user ID'),
           TextFormField(
             key: _formKey,
             decoration: const InputDecoration(
@@ -67,19 +73,27 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
             onPressed: enableNavigation
                 ? () => Navigator.of(
                     context,
-                  ).pushNamed(samsungUserManagementRoute)
+                  ).pushNamed(androidBackgroundStepsRoute)
+                : null,
+            child: const Text('Background steps'),
+          ),
+          FilledButton(
+            onPressed: enableNavigation
+                ? () => Navigator.of(
+                    context,
+                  ).pushNamed(androidUserManagementRoute)
                 : null,
             child: const Text('User management'),
           ),
           FilledButton(
             onPressed: enableNavigation
-                ? () => Navigator.of(context).pushNamed(samsungPermissionsRoute)
+                ? () => Navigator.of(context).pushNamed(androidPermissionsRoute)
                 : null,
             child: const Text('Permissions'),
           ),
           FilledButton(
             onPressed: enableNavigation
-                ? () => Navigator.of(context).pushNamed(samsungSyncRoute)
+                ? () => Navigator.of(context).pushNamed(androidSyncRoute)
                 : null,
             child: const Text('Manually sync health data'),
           ),
@@ -87,7 +101,7 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
             onPressed: enableNavigation
                 ? () => Navigator.of(
                     context,
-                  ).pushNamed(samsungBackgroundSyncRoute)
+                  ).pushNamed(androidBackgroundSyncRoute)
                 : null,
             child: const Text('Background sync'),
           ),
@@ -109,8 +123,8 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
     return null;
   }
 
-  void initialize() async {
-    final configuration = RookConfiguration(
+  void setConfiguration() async {
+    final rookConfiguration = RookConfiguration(
       clientUUID: Secrets.clientUUID,
       secret: Secrets.secret,
       environment: rookEnvironment,
@@ -121,23 +135,27 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
     configurationOutput.clear();
 
     configurationOutput.append('Using configuration:');
-    configurationOutput.append('$configuration');
+    configurationOutput.append('$rookConfiguration');
 
     if (isDebug) {
-      RookSamsung.enableNativeLogs();
+      HCRookConfigurationManager.enableNativeLogs();
     }
+
+    HCRookConfigurationManager.setConfiguration(rookConfiguration);
 
     setState(() {
       configurationOutput.append('Configuration set successfully');
     });
+  }
 
+  void initialize() {
     initializeOutput.clear();
 
     setState(() {
       initializeOutput.append('Initializing...');
     });
 
-    RookSamsung.initRook(configuration)
+    HCRookConfigurationManager.initRook()
         .then((_) {
           setState(() {
             initializeOutput.append('SDK initialized successfully');
@@ -154,7 +172,7 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
   void checkUserIDRegistered() {
     updateUserOutput.clear();
 
-    RookSamsung.getUserID().then((userID) {
+    HCRookConfigurationManager.getUserID().then((userID) {
       if (userID != null) {
         setState(() {
           updateUserOutput.append(
@@ -179,7 +197,7 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
       updateUserOutput.append('Updating userID...');
     });
 
-    RookSamsung.updateUserID(userID!)
+    HCRookConfigurationManager.updateUserID(userID!)
         .then((_) {
           setState(() {
             updateUserOutput.append('userID updated successfully');
@@ -201,7 +219,8 @@ class _SamsungConfigurationState extends State<SamsungConfiguration> {
     });
 
     try {
-      final diagnosticState = await RookSamsung.getDiagnosticState();
+      final diagnosticState =
+          await HCRookConfigurationManager.getDiagnosticState();
 
       diagnosticOutput.appendMultiple([
         "Diagnostic state retrieved",
