+++ lib/sdk/sdk_samsung_health/samsung_user_management.dart
@@ -1,21 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
-import 'package:rook_flutter_sdk/common/console_output.dart';
-import 'package:rook_flutter_sdk/common/widget/scrollable_scaffold.dart';
-import 'package:rook_flutter_sdk/common/widget/section_title.dart';
-import 'package:rook_sdk_health_connect/rook_sdk_health_connect.dart';
+import 'package:rook_flutter_sdk/core/domain/utils/console_output.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/scrollable_scaffold.dart';
+import 'package:rook_flutter_sdk/core/presentation/widget/section_title.dart';
+import 'package:rook_sdk_samsung_health/rook_sdk_samsung_health.dart';
 
-const String androidUserManagementRoute = '/android/user-management';
+const String samsungUserManagementRoute = '/samsung/user-management';
 
-class AndroidUserManagement extends StatefulWidget {
-  const AndroidUserManagement({super.key});
+class SamsungUserManagement extends StatefulWidget {
+  const SamsungUserManagement({super.key});
 
   @override
-  State<AndroidUserManagement> createState() => _AndroidUserManagementState();
+  State<SamsungUserManagement> createState() => _SamsungUserManagementState();
 }
 
-class _AndroidUserManagementState extends State<AndroidUserManagement> {
-  final Logger logger = Logger('AndroidUserManagement');
+class _SamsungUserManagementState extends State<SamsungUserManagement> {
+  final Logger logger = Logger('SamsungUserManagement');
 
   final ConsoleOutput deleteUserOutput = ConsoleOutput();
   final ConsoleOutput syncTimezoneUserOutput = ConsoleOutput();
@@ -24,7 +24,7 @@ class _AndroidUserManagementState extends State<AndroidUserManagement> {
 
   @override
   void initState() {
-    HCRookConfigurationManager.getUserID()
+    RookSamsung.getUserID()
         .then((userID) {
           setState(() {
             this.userID = userID;
@@ -71,7 +71,7 @@ class _AndroidUserManagementState extends State<AndroidUserManagement> {
       syncTimezoneUserOutput.append('Updating user timezone...');
     });
 
-    HCRookConfigurationManager.syncUserTimeZone()
+    RookSamsung.syncUserTimeZone()
         .then((_) {
           setState(() {
             syncTimezoneUserOutput.append('User timezone updated successfully');
@@ -93,7 +93,7 @@ class _AndroidUserManagementState extends State<AndroidUserManagement> {
       deleteUserOutput.append('Deleting user from rook...');
     });
 
-    HCRookConfigurationManager.deleteUserFromRook()
+    RookSamsung.deleteUserFromRook()
         .then((_) {
           setState(() {
             deleteUserOutput.append('User deleted from rook');
